// src/config/interfaces.ts
export type EnvVars = Partial<{
  PORT: number;
  HOST: string;
  APP_NAME: string;
  API_PREFIX: string;
  API_EXCLUDES: boolean;
  PUBLIC_URL: string;
  MORGAN_ENABLED: boolean;
  MORGAN_TYPE: string;
  MONGODB_AUTHSOURCE: string;
  MONGODB_USERNAME: string;
  MONGODB_PASSWORD: string;
  MONGODB_IS_TLS: boolean;
  MONGODB_TLS_KEY: string;
  MONGODB_TLS_INSECURE: boolean;
  MONGODB_DEBUG: boolean;
  DATABASE_URL: string;
  JWT_ISSUER: string;
  JWT_AUDIENCE: string;
  JWT_EXPIRATION: string;
  JWT_PUBLIC_KEY: string;
  JWT_PRIVATE_KEY: string;
  CORS_ENABLED: boolean;
  NODE_ENV: string;
  CORS_ORIGIN: string;
  SECURITY_DELAY: number;
  SECURITY_MAX_TRIES: number;
  SECURITY_CODE_LENGTH: number;
  SECURITY_MAX_SENDS: number;
  SECURITY_TRY_DELAY: number;
  SECURITY_CODE_TTL: number;
  MAILGUN_API_KEY: string;
  MAILGUN_DOMAIN: string;
  MAILGUN_FROM: string;
  MEDIA_STRICT_MODE: boolean;
  LINODE_ENDPOINT: string;
  LINODE_REGION: string;
  LINODE_ACCESS_KEY: string;
  LINODE_SECRET_KEY: string;
  LINODE_BUCKET_URL: string;
  LINODE_BUCKET_NAME: string;
  AUTHORIZED_API_KEY: string;
  AUTHORIZED_TRANSACTION_KEY: string;
  AUTHORIZED_ENVIRONMENT: string;
}>;

export type EnvKeys = keyof EnvVars;